requirements.txt
opencv-python
numpy
pillow
pyttsx3
scikit-learn
matplotlib